+++ CHANGELOG.md
@@ -16,3 +16,4 @@
 
 - Fixes in model
 - Bruno api added
+- Controllers and mappers added